Renames "MyTheme" to "MovLitTheme"
Updates the theme name and associated resources from "MyTheme" to "MovLitTheme" for consistency and clarity.

This includes renaming directories, namespaces, resource files, and updating theme information within the Oqtane CMS.

Also, modifies the Ink module to process all available lines from the story at once.

Adds a default connection string in appsettings.json.
