+++ Huge.MovLit/Client/Themes/MovLitTheme/Containers/Container.razor
@@ -1,4 +1,4 @@
-@namespace Huge.MovLit.MyTheme
+@namespace Huge.MovLit.MovLitTheme
 @inherits ContainerBase
 @inject ISettingService SettingService
