+++ Huge.MovLit/Client/Themes/MovLitTheme/Containers/ContainerSettings.razor
@@ -1,4 +1,4 @@
-@namespace Huge.MovLit.MyTheme
+@namespace Huge.MovLit.MovLitTheme
 @inherits ModuleBase
 @implements Oqtane.Interfaces.ISettingsControl
 @inject ISettingService SettingService
@@ -17,7 +17,7 @@
 </div>
 
 @code {
-    private string resourceType = "Huge.MovLit.MyTheme.ContainerSettings, Huge.MovLit.Client.Oqtane"; // for localization
+    private string resourceType = "Huge.MovLit.MovLitTheme.ContainerSettings, Huge.MovLit.Client.Oqtane"; // for localization
     private string _title = "true";
 
     protected override void OnInitialized()
