+++ Huge.MovLit/Client/Themes/MovLitTheme/ThemeInfo.cs
@@ -3,17 +3,17 @@ using Oqtane.Models;
 using Oqtane.Themes;
 using Oqtane.Shared;
 
-namespace Huge.MovLit.MyTheme
+namespace Huge.MovLit.MovLitTheme
 {
     public class ThemeInfo : ITheme
     {
         public Oqtane.Models.Theme Theme => new Oqtane.Models.Theme
         {
-            Name = "MyTheme",
+            Name = "MovLitTheme",
             Version = "1.0.0",
             PackageName = "Huge.MovLit",
-            ThemeSettingsType = "Huge.MovLit.MyTheme.ThemeSettings, Huge.MovLit.Client.Oqtane",
-            ContainerSettingsType = "Huge.MovLit.MyTheme.ContainerSettings, Huge.MovLit.Client.Oqtane",
+            ThemeSettingsType = "Huge.MovLit.MovLitTheme.ThemeSettings, Huge.MovLit.Client.Oqtane",
+            ContainerSettingsType = "Huge.MovLit.MovLitTheme.ContainerSettings, Huge.MovLit.Client.Oqtane",
             Resources = new List<Resource>()
             {
                 new Stylesheet(Constants.BootstrapStylesheetUrl, Constants.BootstrapStylesheetIntegrity, "anonymous"),
