+++ Huge.MovLit/Client/Themes/MovLitTheme/Themes/Theme.razor
@@ -1,4 +1,4 @@
-@namespace Huge.MovLit.MyTheme
+@namespace Huge.MovLit.MovLitTheme
 @inherits ThemeBase
 @inject ISettingService SettingService
 
@@ -95,7 +95,7 @@
 </main>
 
 @code {
-    public override string Name => "MyTheme";
+    public override string Name => "MovLitTheme";
 
     public override string Panes => PaneNames.Admin + ",Top Full Width,Top 100%,Left 50%,Right 50%,Left 33%,Center 33%,Right 33%,Left Outer 25%,Left Inner 25%,Right Inner 25%,Right Outer 25%,Left 25%,Center 50%,Right 25%,Left Sidebar 66%,Right Sidebar 33%,Left Sidebar 33%,Right Sidebar 66%,Bottom 100%,Bottom Full Width";
