+++ Huge.MovLit/Client/Themes/MovLitTheme/Themes/ThemeSettings.razor
@@ -1,4 +1,4 @@
-@namespace Huge.MovLit.MyTheme
+@namespace Huge.MovLit.MovLitTheme
 @inherits ModuleBase
 @implements Oqtane.Interfaces.ISettingsControl
 @inject ISettingService SettingService
@@ -43,7 +43,7 @@
 
     @code {
         private int pageId = -1;
-        private string resourceType = "Huge.MovLit.MyTheme.ThemeSettings, Huge.MovLit.Client.Oqtane"; // for localization
+        private string resourceType = "Huge.MovLit.MovLitTheme.ThemeSettings, Huge.MovLit.Client.Oqtane"; // for localization
         private string _scope = "page";
         private string _login = "-";
         private string _register = "-";
