+++ Huge.MovLit/Server/appsettings.json
@@ -5,7 +5,7 @@
     "DefaultDBType": "Oqtane.Database.SqlServer.SqlServerDatabase, Oqtane.Server"
   },
   "ConnectionStrings": {
-    "DefaultConnection": ""
+    "DefaultConnection": "Data Source=(LocalDb)\\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\\Oqtane-202509201810.mdf;Initial Catalog=Oqtane-202509201810;Integrated Security=SSPI;Encrypt=false;"
   },
   "Installation": {
     "DefaultAlias": "",
@@ -59,5 +59,9 @@
     "LogLevel": {
       "Default": "Information"
     }
-  }
+  },
+  "InstallationId": "8fa688c3-eb7f-453b-8e58-78b586a2aae5",
+  "InstallationVersion": "6.2.0",
+  "InstallationDate": "202509201811",
+  "PackageRegistryEmail": "host@host.com"
 }
